kinda have it working with an html AND desktop client - avoid generic…

+++ core/src/com/cormicopiastudios/multiplayerpractice/GameEngine/GameMaster.java
@@ -45,7 +45,7 @@ public class GameMaster {
             final PlayerPos posUpdate = new PlayerPos();
             posUpdate.x = pos.x;
             posUpdate.y = pos.y;
-            parent.m.put(Long.valueOf(parent.tid), posUpdate);
+            parent.m.put(Integer.valueOf(parent.tid), posUpdate);
             socket.send(posUpdate);
         }
     }
